+++ pkgs/source_span/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fixed another case in which `FileSpan.union` could throw an exception for
   external implementations of `FileSpan`.
+* Optimize `getLine()` in `SourceFile` when repeatedly called.
 
 # 1.1.4